Issue Id   : #365 Message    : Contract clauses configuration reworks Transcript : The organization has three different types of contracts for employees, each of which has different clauses. This control panel will have a dedicated tab for configuring all three types of contract clauses:
Permanent
Part-time
Probationary

The clauses for each type will be held within collapsible panels that can expand to show the current values of the clauses for that type of contract within rich text input boxes. Administrators can change the clauses for a contract type by providing updated values in the text boxes and clicking a Save button that will become available in the panel once changes are made.